old/requirements.txt
bcolz>=1.1.2
bleach>=2.0.0
certifi>=2016.2.28
cycler>=0.10.0
decorator>=4.1.2
entrypoints>=0.2.3
graphviz>=0.8.2
html5lib>=0.999999999
ipykernel>=4.6.1
ipython>=6.2.0
ipython-genutils>=0.2.0
ipywidgets>=7.0.1
isoweek>=1.3.3
jedi>=0.10.2
Jinja2>=2.9.6
jsonschema>=2.6.0
jupyter>=1.0.0
jupyter-client>=5.1.0
jupyter-console>=5.2.0
jupyter-core>=4.3.0
MarkupSafe>=1.0
matplotlib>=2.0.2
mistune>=0.7.4
nbconvert>=5.3.1
nbformat>=4.4.0
notebook>=5.1.0
numpy>=1.13.1
olefile>=0.44
opencv-python>=3.3.0.10
pandas>=0.20.3
pandas_summary>=0.0.5
pandocfilters>=1.4.2
pexpect>=4.2.1
pickleshare>=0.7.4
Pillow>=4.2.1
prompt-toolkit>=1.0.15,<2.0.0
ptyprocess>=0.5.2
Pygments>=2.2.0
pyparsing>=2.2.0
pytest>=3.5.0
pytest-cov>=2.5.1
python-dateutil>=2.6.1
pytz>=2017.2
PyYAML>=3.12
pyzmq>=16.0.2
qtconsole>=4.3.1
scipy>=0.19.1
seaborn>=0.8.1
simplegeneric>=0.8.1
six>=1.11.0
sklearn_pandas>=1.6.0
terminado>=0.6
testpath>=0.3.1
torch<0.4
torchtext>=0.2.3
torchvision>=0.2.0
tornado>=4.5.2,<5
tqdm>=4.15.0
traitlets>=4.3.2
wcwidth>=0.1.7
webencodings>=0.5.1
widgetsnbextension>=3.0.3
spacy>=2.0.11
testfixtures>=6.2.0
